refactor: moved off from builder pattern for params as it made code unnecesarly complicated

+++ src/use_cases/group/revoke_permission.rs
@@ -23,7 +23,7 @@ impl GroupsUseCase {
         };
         use crate::use_cases::UsersUseCase;
         use crate::errors::use_case::GroupRevokePermissionError as Error;
-        use crate::params::repository::GroupPermissionDeleteParamsBuilder as ParamsBuilder;
+        use crate::params::repository::GroupPermissionDeleteParams as Params;
 
         let mut client = client.acquire()
             .await
@@ -31,16 +31,9 @@ impl GroupsUseCase {
         
         // authorize
         {
-            use crate::params::use_case::UserAuthorizeParamsBuilder as ParamsBuilder;
-
-            let params = ParamsBuilder::new()
-                .set_token(params.token)
-                .set_encryption_key(params.encryption_key)
-                .set_permission_name(String::from("authios:all"))
-                .build()
-                .unwrap();
+            use crate::params::use_case::UserAuthorizeParams as Params;
             
-            match UsersUseCase::authorize(params, &mut *client).await {
+            match UsersUseCase::authorize(Params { token: params.token, encryption_key: params.encryption_key, permission_name: String::from("authios:all") }, &mut *client).await {
                 Ok(true) => (),
                 _ => return Err(Error::Unauthorized)
             };
@@ -48,14 +41,9 @@ impl GroupsUseCase {
         
         // check if the permission exist
         {
-            use crate::params::repository::PermissionRetrieveParamsBuilder as ParamsBuilder;
-            
-            let params = ParamsBuilder::new()
-                .set_name(params.permission_name.clone())
-                .build()
-                .unwrap();
+            use crate::params::repository::PermissionRetrieveParams as Params;
 
-            let _ = PermissionsRepository::retrieve(params, &mut *client)
+            let _ = PermissionsRepository::retrieve(Params { name: params.permission_name.clone() }, &mut *client)
                 .await
                 .map_err(|_| Error::PermissionNotFound)?;
         }
@@ -64,14 +52,9 @@ impl GroupsUseCase {
 
         // check if the group exist
         let group = {
-            use crate::params::repository::GroupRetrieveParamsBuilder as ParamsBuilder;
-            
-            let params = ParamsBuilder::new()
-                .set_name(params.group_name.clone())
-                .build()
-                .unwrap();
+            use crate::params::repository::GroupRetrieveParams as Params;
 
-            let group = GroupsRepository::retrieve(params, &mut *client)
+            let group = GroupsRepository::retrieve(Params { name: params.group_name.clone() }, &mut *client)
                 .await
                 .map_err(|_| Error::GroupNotFound)?;
 
@@ -83,17 +66,9 @@ impl GroupsUseCase {
         // check if the permission is added
         if !group.permissions.contains(&params.permission_name) {
             return Err(Error::NotAddedYet);
-        }
-
+        }        
         
-
-        let params = ParamsBuilder::new()
-            .set_group_name(params.group_name)
-            .set_permission_name(params.permission_name)
-            .build()
-            .unwrap();
-        
-        let _ = GroupPermissionsRepository::delete(params, &mut *client)
+        let _ = GroupPermissionsRepository::delete(Params { group_name: params.group_name, permission_name: params.permission_name }, &mut *client)
             .await;
         
         return Ok(());
